Added #10. Communications, adjusted #6. Management
Management: Now anyone on the committee except the President may hold up to two positions. Previously, only the VUWSA rep could hold a second position on the committee
Communications: all conversations originating in contact@vec.ac.nz must be contained in contact@vec.ac.nz in the interest of transparency.
Syntax corrections and clarifications (eg. clarifying that the clause which may NOT be changed is the Interpretation... clause, clarifying that the Cultural Council Representative is otherwise known as VUWSA rep)